+++ src/Namespace.js
@@ -1,27 +1,32 @@
 var Namespace = Object.create(Object.prototype, {
 
   create: {
-    value: function (name) {
+    value: function (name, items) {
       if (name) {
         var fragments = name.split('.'),
-            parent, 
+            ns, 
             position = 0;
         if (Module.exports) {
-          parent = Module.exports;
           if (fragments[0] !== Module.exportsName) {
             throw new TypeException("Module's name is invalid. Expected '" + fragments[0] + "', but found '" 
               + Module.exportsName + "'.");
           }
+          ns = Module.exports;
           position = 1;
         } else {
-          parent = global;
+          ns = global;
         }
         for (var i = position; i < fragments.length; i++) {
-          if (!(fragments[i] in parent)) {
-            parent[fragments[i]] = {}; 
+          if (!(fragments[i] in ns)) {
+            ns[fragments[i]] = {}; 
           }
-         parent = parent[fragments[i]]; 
-        }        
+          ns = ns[fragments[i]]; 
+        }
+        if (items) {
+          for (var propertyName in items) {
+            ns[propertyName] = items[propertyName];
+          }
+        }
       } else {
         throw new TypeException("Namespace is not set");
       }