+++ tests/browser/scripts/class_within_namespace.js
@@ -1,0 +1,20 @@
+$run('Class within namespace', function () {
+
+  'use strict';
+
+  $namespace('myapp.visual', {
+    Shape: $class({
+      __construct__: function () {
+
+      },
+      draw: function () {
+
+      }
+    })
+  });
+
+  $assert(myapp.visual.Shape);
+
+  $complete('Class within namespace');
+
+});